move debug logging to separate file
