+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
 	"compilerOptions": {
 		"module": "commonjs",
-		"target": "es6",
+		"target": "es2020",
 		"outDir": "./out",
 		"lib": ["ES2022", "DOM"], // DOM is Temporary work around for node-fetch blob issue https://github.com/node-fetch/node-fetch/issues/1617
 		"sourceMap": true,
